+++ CHANGELOG.md
@@ -4,16 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.0] - 2022-01-14
+
+## [1.3.0] - 2022-01-17
 ### Added
+- Explanation of how to treat .csv input in troubleshooting
 - New instructions in the README.md to set a more flexible write access to the user created `tmp/` folder.
 
 ### Changed
 - DSL-1 converted to DSL-2
 - Updated nextflow in Dockerfile to nextflow-21.12.1-edge-all
+- Improved `docs/`, but still work to be done
 
 ### Fixed
 - Now possible to run from anywhere, not only from within the clensumstats repo
+- Now DIRECTION is skipped by the awk numeric test
 
 ## [1.2.0] - 2021-12-08
 ### Added